backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
qdrant-client
sentence-transformers
redis
tavily-python
transformers
torch
pillow
sse-starlette
python-dotenv
pdfplumber
langchain-openai
langchain-huggingface
pymupdf
firebase-admin
langchain-qdrant